Moved Display & System setting to frontend protobuf
* Moved Display & System setting to frontend protobuf
